refactor: 문서테스트 패키지 명 변경
